Add support to let pages decide if they have finished to render
A function isRendering can be defined on the object PhantomJSPrinting on
the code of the webpage to export. This function should return true if
the page knows that it is yet being rendering. In this case, it won’t be
exported to pdf yet and a setInterval will be started with a time
calculated as the rendering_time passed to Shrimp divided by 5. After
this time, the page will be asked again about the rendering status,
repeating this process until it return false. In that moment the
exportation to pdf will be started.